[dom] Separate HTML from base element macro.
Also relocates macros to single place.
