+++ dom/src/elements/text_semantics.rs
@@ -11,7 +11,7 @@ use crate::{
 };
 use augdom::event;
 
-element! {
+html_element! {
     /// The [HTML `<a>` element (or *anchor* element)][mdn], along with its href attribute, creates
     /// a hyperlink to other web pages, files, locations within the same page, email addresses, or
     /// any other URL.
@@ -20,7 +20,7 @@ element! {
     a -> Anchor
 }
 
-element! {
+html_element! {
     /// The [HTML Abbreviation element (`<abbr>`)][mdn] represents an abbreviation or acronym; the
     /// optional [`title`][title] attribute can provide an expansion or description for the
     /// abbreviation.
@@ -30,7 +30,7 @@ element! {
     abbr -> Abbreviation
 }
 
-element! {
+html_element! {
     /// The [HTML Bring Attention To element (`<b>`)][mdn] is used to draw the reader's attention to
     /// the element's contents, which are not otherwise granted special importance.
     ///
@@ -38,7 +38,7 @@ element! {
     b -> BringAttentionTo
 }
 
-element! {
+html_element! {
     /// The [HTML Bidirectional Isolate element (`<bdi>`)][mdn] tells the browser's bidirectional
     /// algorithm to treat the text it contains in isolation from its surrounding text.
     ///
@@ -46,7 +46,7 @@ element! {
     bdi -> BidirectionalIsolate
 }
 
-element! {
+html_element! {
     /// The [HTML Bidirectional Text Override element (`<bdo>`)][mdn] overrides the current
     /// directionality of text, so that the text within is rendered in a different direction.
     ///
@@ -54,7 +54,7 @@ element! {
     bdo -> BidirectionalOverride
 }
 
-element! {
+html_element! {
     /// The [HTML `<br>` element][mdn] produces a line break in text (carriage-return). It is useful
     /// for writing a poem or an address, where the division of lines is significant.
     ///
@@ -62,7 +62,7 @@ element! {
     br -> LineBreak
 }
 
-element! {
+html_element! {
     /// The [HTML Citation element (`<cite>`)][mdn] is used to describe a reference to a cited
     /// creative work, and must include the title of that work.
     ///
@@ -70,7 +70,7 @@ element! {
     cite -> Citation
 }
 
-element! {
+html_element! {
     /// The [HTML `<code>` element][mdn] displays its contents styled in a fashion intended to
     /// indicate that the text is a short fragment of computer code.
     ///
@@ -78,7 +78,7 @@ element! {
     code -> Code
 }
 
-element! {
+html_element! {
     /// The [HTML `<data>` element][mdn] links a given content with a machine-readable translation.
     /// If the content is time- or date-related, the [`<time>`][time] element must be used.
     ///
@@ -87,7 +87,7 @@ element! {
     data -> Data
 }
 
-element! {
+html_element! {
     /// The [HTML Definition element (`<dfn>`)][mdn] is used to indicate the term being defined
     /// within the context of a definition phrase or sentence.
     ///
@@ -95,7 +95,7 @@ element! {
     dfn -> Definition
 }
 
-element! {
+html_element! {
     /// The [HTML `<em>` element][mdn] marks text that has stress emphasis. The `<em>` element can
     /// be nested, with each level of nesting indicating a greater degree of emphasis.
     ///
@@ -103,7 +103,7 @@ element! {
     em -> Emphasis
 }
 
-element! {
+html_element! {
     /// The [HTML `<i>` element][mdn] represents a range of text that is set off from the normal
     /// text for some reason. Some examples include technical terms, foreign language phrases, or
     /// fictional character thoughts. It is typically displayed in italic type.
@@ -112,7 +112,7 @@ element! {
     i -> Italic
 }
 
-element! {
+html_element! {
     /// The [HTML Keyboard Input element (`<kbd>`)][mdn] represents a span of inline text denoting
     /// textual user input from a keyboard, voice input, or any other text entry device.
     ///
@@ -120,7 +120,7 @@ element! {
     kbd -> KeyboardInput
 }
 
-element! {
+html_element! {
     /// The [HTML Mark Text element (`<mark>`)][mdn] represents text which is marked or highlighted
     /// for reference or notation purposes, due to the marked passage's relevance or importance in
     /// the enclosing context.
@@ -129,7 +129,7 @@ element! {
     mark -> Mark
 }
 
-element! {
+html_element! {
     /// The [HTML `<q>` element][mdn]  indicates that the enclosed text is a short inline quotation.
     /// Most modern browsers implement this by surrounding the text in quotation marks.
     ///
@@ -137,7 +137,7 @@ element! {
     q -> Quotation
 }
 
-element! {
+html_element! {
     /// The [HTML Ruby Base (`<rb>`) element][mdn] is used to delimit the base text component of
     /// a [`<ruby>`][ruby] annotation, i.e. the text that is being annotated.
     ///
@@ -146,7 +146,7 @@ element! {
     rb -> RubyBase
 }
 
-element! {
+html_element! {
     /// The [HTML Ruby Fallback Parenthesis (`<rp>`) element][mdn] is used to provide fall-back
     /// parentheses for browsers that do not support display of ruby annotations using the
     /// [`<ruby>`][ruby] element.
@@ -156,7 +156,7 @@ element! {
     rp -> RubyFallbackParens
 }
 
-element! {
+html_element! {
     /// The [HTML Ruby Text (`<rt>`) element][mdn] specifies the ruby text component of a ruby
     /// annotation, which is used to provide pronunciation, translation, or transliteration
     /// information for East Asian typography. The `<rt>` element must always be contained within a
@@ -167,7 +167,7 @@ element! {
     rt -> RubyText
 }
 
-element! {
+html_element! {
     /// The [HTML Ruby Text Container (`<rtc>`) element][mdn] embraces semantic annotations of
     /// characters presented in a ruby of [`<rb>`][rb] elements used inside of [`<ruby>`][ruby]
     /// element. [`<rb>`][rb] elements can have both pronunciation ([`<rt>`][rt]) and semantic
@@ -181,7 +181,7 @@ element! {
     rtc -> RubyTextContainer
 }
 
-element! {
+html_element! {
     /// The [HTML `<ruby>` element][mdn] represents a ruby annotation. Ruby annotations are for
     /// showing pronunciation of East Asian characters.
     ///
@@ -189,7 +189,7 @@ element! {
     ruby -> Ruby
 }
 
-element! {
+html_element! {
     /// The [HTML `<s>` element][mdn] renders text with a strikethrough, or a line through it. Use
     /// the `<s>` element to represent things that are no longer relevant or no longer accurate.
     /// However, `<s>` is not appropriate when indicating document edits; for that, use the
@@ -201,7 +201,7 @@ element! {
     s -> Strikethrough
 }
 
-element! {
+html_element! {
     /// The [HTML Sample Element (`<samp>`)][mdn] is used to enclose inline text which represents
     /// sample (or quoted) output from a computer program.
     ///
@@ -209,7 +209,7 @@ element! {
     samp -> Sample
 }
 
-element! {
+html_element! {
     /// The [HTML `<small>` element][mdn] represents side-comments and small print, like copyright
     /// and legal text, independent of its styled presentation. By default, it renders text within
     /// it one font-size small, such as from `small` to `x-small`.
@@ -218,7 +218,7 @@ element! {
     small -> Small
 }
 
-element! {
+html_element! {
     /// The [HTML `<span>` element][mdn] is a generic inline container for phrasing content, which
     /// does not inherently represent anything. It can be used to group elements for styling
     /// purposes (using the [`class`][class] or [`id`][id] attributes), or because they share
@@ -231,7 +231,7 @@ element! {
     span -> Span
 }
 
-element! {
+html_element! {
     /// The [HTML Strong Importance Element (`<strong>`)][mdn] indicates that its contents have
     /// strong importance, seriousness, or urgency. Browsers typically render the contents in bold
     /// type.
@@ -240,7 +240,7 @@ element! {
     strong -> Strong
 }
 
-element! {
+html_element! {
     /// The [HTML Subscript element (`<sub>`)][mdn] specifies inline text which should be displayed
     /// as subscript for solely typographical reasons.
     ///
@@ -248,7 +248,7 @@ element! {
     sub -> Subscript
 }
 
-element! {
+html_element! {
     /// The [HTML Superscript element (`<sup>`)][mdn] specifies inline text which is to be displayed
     /// as superscript for solely typographical reasons.
     ///
@@ -256,14 +256,14 @@ element! {
     sup -> Superscript
 }
 
-element! {
+html_element! {
     /// The [HTML `<time>` element][mdn] represents a specific period in time.
     ///
     /// [mdn]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/time
     time -> Time
 }
 
-element! {
+html_element! {
     /// The [HTML Unarticulated Annotation Element (`<u>`)][mdn] represents a span of inline text
     /// which should be rendered in a way that indicates that it has a non-textual annotation.
     ///
@@ -271,7 +271,7 @@ element! {
     u -> Unarticulated
 }
 
-element! {
+html_element! {
     /// The [HTML Variable element (`<var>`)][mdn] represents the name of a variable in a
     /// mathematical expression or a programming context.
     ///
@@ -279,7 +279,7 @@ element! {
     var -> Variable
 }
 
-element! {
+html_element! {
     /// The [HTML `<wbr>` element][mdn] represents a word break opportunity—a position within text
     /// where the browser may optionally break a line, though its line-breaking rules would not
     /// otherwise create a break at that location.
@@ -288,7 +288,7 @@ element! {
     wbr -> WordBreakOpportunity
 }
 
-element! {
+html_element! {
     /// The [HTML `<del>` element][mdn] represents a range of text that has been deleted from a
     /// document.
     ///
@@ -296,7 +296,7 @@ element! {
     del -> Deleted
 }
 
-element! {
+html_element! {
     /// The [HTML `<ins>` element][mdn] represents a range of text that has been added to a
     /// document.
     ///
